BaseballGame with OOP
