+++ haeeun/BaseballGame/src/main/main.java
@@ -1,0 +1,8 @@
+package main;
+
+public class main {
+    public static void main(String[] args) {
+        Baseball baseball = new Baseball();
+        baseball.Game();
+    }
+}
